+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Unreleased
 
+* Add `ShortcutText!` widget for displaying keyboard shortcuts.
+    - Add `zng-wgt-shortcut` crate and `zng::shortcut_text` module.
+    - Also provides localization for key names. !!: TODO
+    - Widget used in menu command shortcut styles. !!: TODO
+
 * Improve `zng::fs_watcher`, resolve config desync issues.
 
 * Fix `Markdown!` whitespace merging.